+++ src/core/Application.h
@@ -1,6 +1,10 @@
 #pragma once
 
+#include <memory>
+#include <vector>
+
 class Window;
+class EditorPanel;
 
 class Application
 {
@@ -15,4 +19,5 @@ public:
 private:
     Window *m_window;
     bool m_running;
+    std::vector<std::shared_ptr<EditorPanel>> m_panels;
 };