+++ src/editor/EditorPanel.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class EditorPanel
+{
+public:
+    virtual ~EditorPanel() = default;
+    virtual void OnImGuiRender(float dt) = 0;
+};